+++ MillenniumEye/model/Card.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace MillenniumEye.model
+{
+    public class Card
+    {
+        [JsonProperty("data", NullValueHandling = NullValueHandling.Ignore)]
+        public List<Datum>? Data { get; set; }
+    }
+}
